Simplify SVGLength::CloneForAnimation
If SetValueAsString(...) fails it will not modify |value_|, which means
it will remain as '0' which is the initial value, so no need to reset
the value. Also, use the constructor that takes a SVGLengthMode.

Remove the copy constructor because the only user can use the one that
takes a value as an argument instead.

Change-Id: Iad3f01d162d64e47a1f6da542c3b26859d74d317
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174400
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764368}