Migrate FrameMsg_Stop from the legacy IPC to blink::mojom::LocalFrame
This CL moves this message out of frame_messages.h and adds a new
method in blink::mojom::LocalFrame, implemented by blink::LocalFrame,
to request the renderer to stop a load if there's any in progress.

Unfortunately, due to a dependency in PepperURLLoaderHost::Close(),
we can not get rid entirely of having a StopLoading() method in
Blink's public API. Therefore, a WebLocalFrame::StopLoading() still
needs to be added to satisfy that dependency, even though calls to
this StopLoading() method from the browser process will now be routed
directly to Blink.

Bug: 1064344
Change-Id: I3030d136c07715aae8c52854b901e9b7f43bb645
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170232
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764399}