[Web Payments] Payment apps with missing icons are not JIT installable.
After this cl a service worker based payment app will not be available
for JIT installation if the crawler cannot download a non-empty icon for
it.
This change prevents payment apps from getting stuck in a broken
installed state with missing icon. This is particularly important since
currently Chrome does not support refetching icons for already installed
payment apps, and apps with missing icons are not preselectable (meaning
that the payment sheet will always be shown for the user to explicitly
choose the app to proceed with.)

Change-Id: I4d7bf0bf199972c584b7c191b34dfe95d2d4c29f
Bug: 1069010
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173007
Commit-Queue: Sahel Sharify <sahel@chromium.org>
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764404}