Add tooltip string for extensions disabled due to parent permissions
If the parent disables the "Permissions for sites, apps and
extensions" toggle, and the supervised user has an extension that
still requires custodian approval or has permissions increases, then
that extension should be force-disabled on the chrome://extensions
page with a tool-tip explanation. This CL commits the required string
for the tool-tip before the M84 feature freeze on May 1, 2020, so that
the translation team has adequate time to translate the string.

Bug: 1068655
Change-Id: I7e7efb4e28d66488d744fcc255016e83cb482faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173959
Reviewed-by: Ken Rockot <rockot@google.com>
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764407}