[WebLayer] Control site settings help menu option with SiteSettingsClient.
This CL introduces a SiteSettingsHelpClient interface that pulls in the
existing help related SiteSettingsClient methods, and introduces a new
method to control the visibility of help menu items. The help menu
behavior is currently Chrome-specific, so this provides us a mechanism
to hide the options for non-Chrome embedders.

Bug: 1058595
Change-Id: I4c3db35655fa1f237e019baced141120de1ce44a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164028
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Natalie Chouinard <chouinard@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764418}