[FlexNG] Don't miss the intrinsic-sizes cache for tables.
This was a bug when this logic was first introduced. Effectively we
always needed miss the intrinsic-sizes cache to pass:
external/wpt/css/vendor-imports/mozilla/mozilla-central-reftests/flexbox/flexbox-align-self-horiz-001-table.xhtml

This was incorrect, this was hiding the bug. When we compute the min/max
sizes from legacy, tables return their preferred inline-sizes, where-as
non-tables return their intrinsic inline-sizes.

If the node was a table - we should be returning the preferred inline-sizes,
vs. the intrinsic inline-sizes.

Bug: 845235
Change-Id: Ibf2ce053fcdb2898724644aa1cc847f0d5c56b22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174789
Reviewed-by: David Grogan <dgrogan@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764427}