[ua-ch] Avoid leaking android model as part of platformVersion
GetAndroidOSInfo is currently also appending the model information to
the result, which caused a leak of that model information in UA-CH
platformVersion, which relied in it.

This CL refactors that function to make sure it doesn't do that in
instances where it's not desired. It also adds unit tests to the
User-Agent string, to make sure it's not broken by this change, and to
avoid issues like https://crbug.com/1061132 in the future.

Bug: 1072460
Change-Id: Id788a51f2cdf01e84c45e53885d3e8e1edd9c815
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166164
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Aaron Tagliaboschi <aarontag@chromium.org>
Commit-Queue: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764428}