Web UI: Allow invalid chrome:// file requests to fail.
Don't set a default resource for Web UIs in
SetupWebUIDataSource()/SetupBundledWebUIDataSource(). Instead, set the
resource passed in as default as the resource for the empty path.

In WebUIDataSource, for sources without a default resource, return
the empty path resource (if it exists) if the request is for a path
that doesn't have a mime type and isn't HTML.

Bug: None
Change-Id: I92875670d790231e0cfb6fbb254bcfb7ffac1a6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173380
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764439}