Use gpuBenchmarking in fast/events/dispatch-mouse-events-to-window-always
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in scroll tests in fast/scroll-behavior/smooth-scroll/.

Bug: 1047176
Change-Id: I8726ad59de20ce79f5eae9ace26370a5bb95507f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173144
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764467}