weblayer: converts FeatureListCreator to use local-state
FeatureListCreator was creating it's own pref store, it should
be using local-state instead. This is an incompatible change,
but it was only finch state stored in these prefs and should be
fine to recreate.

This patch makes it so we create local-state earlier, and roughly
matches what happens in chrome. BrowserProcess continues to own
local-state, which means local-state is injected into BrowserProcess
(again, the same thing happens in chrome).

BUG=1073068
TEST=none

Change-Id: Ie09e698478f66807e2896db443bbfba1ade4c987
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173392
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764508}