Check target allowlist in Target Embedding heuristic.
This CL adds support for checking the target allowlist before flagging
a domain for target embedding. In target embedding heuristic, when a
domain is targeted, we suggest the targeted domain to the user for
redirection (e.g. google.com in google.com-foo.com). This is also the
case even if there is a targeted subdomain in the target embedding
domain (e.g. google.com is suggested for scholar-google.com-foo.com).
However, if the subdomain is allowlisted and the domain itself is not
allowlisted, we will make sure to not flag the target embedding domain.
That is, if google.com is not allowlisted, and scholar.google.com is
allowlisted, scholar.google.com.foo.com will not be flagged as a target
embedding domain.

Bug: 1075659,1075739
Change-Id: I5287bab80571e36b8bd120e814a0eda6347119b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173826
Commit-Queue: Behnood Momenzadeh <behnoodm@google.com>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764519}