Update mixed content settings to use map of RFH -> setting
Currently mixed content settings is saved in MixedContentSettingsTabHelper
and updated manually on each main-frame navigation, resetting it when we
switch SiteInstances. After we navigated away, it is not possible to know
what a previous page's mixed content settings were because the value
might already be overwritten. This means it's not possible to restore the
settings of a page when it gets restored from the back-forward cache,
which is why currently we disable BFCache in this case.

We are changing it to use a map of RFH -> setting instead so that we can
restore it (as the RFH will not be deleted), and in the near
future when RenderDocument is implemented (RFH will change on every
navigation to a different document , we will save mixed content settings
as a RenderDocumentHostUserData which means it will have the same
lifetime as a document, which means it will stay around when a page we
navigated away from is put into bfcache, and the settings can be restored
when the page is restored from the bfcache as well.

Bug: 1071232
Change-Id: I545cbf80c1be33149d51e392e712de73af75a6a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142073
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764540}