Fix and improve gdb+vscode integration
The `libcxx-pretty-printers` library contained issues where pointers
could not be expanded, due to the lack of `children` impl.

Switching to the pretty printers included with the llvm libcxx project
fixes these issues. It will also ensure that the pretty printers are
upgraded in step with the libcxx revision.

This also updates the included vscode settings with sensible defaults,
used to enable pretty printing and load the Chromium and Blink gdb
configurations.

Change-Id: I3cbdb778f245618866f3f5c6f217a72bc97e26a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2166230
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764542}