Exposing GetComputedValue to WebElement
The change adds exposure of CSS property values are part of the
ComputedStyle to WebElement. The functionality is usually exposed
through the Window object in WebIDL as window.getComputedStyle(element).
Historically this was done in hope of supporting multi-view
presentations and having Window be the view identifier. This hasn't gone
anywhere yet.

Exposing ComputedStyle requires all of CSSComputedStyleDeclaration which
is a pretty large interfaces, thus, for now we are exposing computed
property values as strings directly to WebElement.

This enables use of computed style property values by public component
usage through /public/web interfaces.

Change-Id: Ia926190abd0e3dfba4c848607735e18430d02817
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173400
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Rossen Atanassov <ratan@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#764661}