Fix Resize Observer device-pixel-content-box on Mac
Device pixel content box on Macbooks does not work as expected.
Currently, the values returned are not in device pixels, but rather CSS
pixels. This occurs because the use-zoom-for-dsf is not enabled at the
moment.

This issue is fixed by multiplying the snapped device pixel sizes by
the device scale factor.

This does not fix the following issue where use-zoom-for-dsf = false:
An element, for example canvas, has odd dimensions like 75x75 and
the device scale factor = 1.5. As actual pixel snapping occurs on
the 75x75 dimensions before it is scaled, we get dimensions that
are different from 113x113. However, use-zoom-for-dsf = false is
relevant for Macbooks, which have been found to have integer dsf's.


Bug: 1073700
Change-Id: I3e3cc7e2f47f8b51518737b14b8be4829e1bacd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2167428
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764690}