Add comments on when gpuBenchmarking finishes a gesture
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. We should replace
eventSender.continuousMouseScrollBy with gpuBenchmarking.smoothScrollBy
in scroll tests in fast/scroll-behavior/smooth-scroll/.

Bug: 1047176
Change-Id: Ifcf2392aeb427a57a1958437613f0ca3d8ab2205
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176151
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764701}