Move instance ID unregistration to Shutdown
This CL fixes two problems with the BinaryFCMService's handling of
unregistration. First, on destruction the GCMDriver may have been
destroyed, leading to a UaF. This is solved by moving the
unregistration to the shutdown pass.

Secondly, the BinaryUploadService was unregistering the wrong tokens.
This was fixed and tests were added to ensure we unregister the FCM
InstanceID.

Bug: 1068732
Change-Id: Ia72c62ef24b8600d1853b1f2232f11028b31b760
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2144839
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Bettina Dea <bdea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764707}