fido: improve macOS platform authenticator availability checks
The macOS Touch ID platform authenticator was previously considered
available whenever the LocalAuthentication framework indicated that it
could authenticate the local user via biometrics (Touch ID or,
hypothetically, Face ID). For macOS, this is generally the case when
 - the device has Touch ID (which implies it also has a secure enclave),
 - Touch ID has at least one fingerprint enrolled,
 - biometrics are not temporarily "locked out" (see crbug/1045646),
 - and the display lid is open.

When the Touch ID authenticator _is_ available, we allow the local user
to authenticate using either their password or Touch ID.  So having it
be _unavailable_ just because we won't be able to authenticate via Touch
ID specifically, even though the password would still work fine, is
unncecessary and annoying.

Instead, change this to test that
 - the local user can be authenticated by any of Touch ID/Face ID, their
   Apple Watch, or their device passcode,
 - and the device has a secure enclave.

Fixed: 1045646, 1002526
Change-Id: I9d816686a6e846e418187f359268300fad79a9d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174968
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#764720}