[WebLayer] GetDarkTheme uses embedder's app context.
WebLayer wraps the application context passed in, in a
ContextThemeWrapper so we can load resources dynamically. However, this
means that the resources we pick up are not the app's resources, rather
the resources from this wrapped context.

To determine whether dark theme is enabled, we're using the wrapped
context. Using this context, we won't get, for example, any configuration
overrides applied by the app (like forcing dark mode on when it's not enabled
system wide) or anything done by compat libraries in AndroidX on older OS
versions.

This CL saves the embedder's app context, and uses that to determine
whether dark theme is enabled. It also updates the embedder app context
whenever appropriate.

Bug: 1074278, 1077025
Change-Id: I53e5ee9d484ee26ca473678c60e88da164e5790b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174203
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764722}