Remove web-test-specific code for focusing.
Drops the SetFocusAndActivate() test code from RenderViewImpl and the
plumbing over to it. Instead, pretend to receive the focus/activate
IPCs that the browser sends in production.

window.focus() on another window goes through the browser process in
production, but had a synchronous override in web tests, that allowed
tests to pass even though we should need to wait to see the focus occur
asynchronously in chrome.

It also denies the request unless there is a user action present, but in
web tests we dropped that requirement. This CL removes the web test
override there, and adds user gestures to the tests that are performing
a window.focus() on another window.

Drop the focus of the main frame when navigating to about:blank between
tests. The browser already sends a Focus() IPC to the main frame of the
about:blank at the start of a test. This was causing the test windows to
lie about being focused - they would all claim to be focused! It caused
an error in one sensor test, where it runs code in the secondary
inspector window that expects the window to be focused. So a mojom
method is added to mojom::WebTestControlHost to move focus to that
window for the test.

Remove all the other extraneous focus-moving that was happening in the
WebTestControlHost and friends browser code. That code was moving focus
on Linux but does nothing on Mac, because on Mac headless mode the
RenderWidgetHostView::Focus() method asks the OS to focus a window that
isn't visible so it does nothing.

Rename the testRunner.setWindowIsKey(bool) to setWindowFocus(bool) and
remove comments explaining that the method is actually setting window
focus. It would be nice to remove this entirely and use a separate
window to move focus around, but that doesn't work for some tests, which
is a bit interesting but not addressed by this CL.

Drop the "previously focused window" tracking in TestRunner, which would
keep around invalid (UAF-prone) pointers. Instead, iterate through the
windows (actually, RenderViews) and blur the focused one there.

The http/tests/serviceworker/clients-openwindow.html test expectations
are adjusted as the renderer no longer marks the main frame as focused
on navigation, so the service worker no longer reports the opened
window as focused.

Filed bug 1074482 and bug 1076129 for WPT tests that fail after this
change. The former because it's either/both expecting run_web_tests.py
to focus newly recreated windows and we don't, and it was calling
window.focus() of another window, which had test-only synchronous code
which we have removed. The latter because it seems to have a bug in how
its written that this CL has uncovered.

R=nasko@chromium.org

Bug: 866140, 1069111, 1074482, 1076129
Change-Id: I9d7ed8f9c41a75fc8afaf6bb0bfc3e5772e1297c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2165248
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764725}