cros: Re-enable use-transforms option for BoundsAnimator.
A fix crrev.com/c/2165680 was landed to fix the original issue. But to
be safe just disable for merge back (also 1 or 2 patches were landed
in between).

This patch re-enables use of the transforms option with one change;
before animating, if bounds on the view are set, explicitly stop
animating the view first as modifiying bounds partway through an
onging animation does not work so well with transforms as the bounds
need to be calculated on animation end/cancel.

Test: none
Bug: 1076153
Change-Id: I3e64f203fc3f3a8bdae89ec2b5a13cf27a1bfa95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2174827
Reviewed-by: Mitsuru Oshima  (Slow) <oshima@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764746}