caBLEv2: once the first client makes a GATT write, reject all other clients
Since the authenticator is designed to handle a single desktop client
only, make it ignore other BLE peers beyond the first one that
establishes a session.

It's not entirely clear at what point a client should be considered
established. The first GATT connection event is probably too unspecific:
GATT connections are not to a particular service, so clients may connect
to interact with other GATT services, and we don't want that to
preclude actual caBLE clients from establishing themselves.

So interaction with the FIDO GATT service is probably a better choice.
Writing to the fidoControlPoint characteristic seems like a good fit in
particular since that triggers initialization of the Client struct on
the C++ side already.

Bug: 1002262
Change-Id: I6c694b067415ab5f0b2f3c286ca50643c584d989
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173739
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764791}