Use DISPLAY priority TaskRunner for silent sink suspension.
Unfortunately the interaction of WebAudio's pull model and MediaStream's
push model ends up requiring a high priority thread to pump WebAudio. In
most cases the AudioOutputDevice provides this thread, but in the MS
case audio is redirected to the MS and only silent audio is delivered to
the actual output device. After 30s the SilentSinkSuspender kicks in to
avoid wasting power on an unused output device; which can be up to 25%
improvement.

SilentSinkSuspender runs on the media thread though, which in addition
to being normal priority, is generally running tasks which may take
longer than the audio time slice (~2ms in the worst on macOS) to run.
As such we can get glitches in the audio flowing out of the MS.

As such, lets give the SilentSinkSuspender a display priority thread and
see if that's sufficient to allow us to keep the power gains as well as
solve the glitching issues.

R=hongchan

Bug: 766198, 1030950
Change-Id: Id1f8b9fef052490fcd15b6b2e923b42e80f0ed2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2176259
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#764816}