Added new method 'GetChildRoot' to track child roots.
This allows us to remove the tree search when looking for the web root
for date popups.
We were previously indirectly doing a tree search on every call to
AXPlatformNodeWin::Navigate when a popup was open. This change will
mitigate the perf impact of:
LegacyRenderWidgetHostHWND::GetOrCreateBrowserAccessibilityRoot

Updated input-color-with-popup-open-expected-uia-win.txt to include
a pane. This was previously missed as we relied on the
one shot accessibilty tree search to find a popup button, which
does not exist in this scenario.

Bug: 1057185
Change-Id: I54beb54132d93f88768f12427d40d6b097566640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076412
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#764821}