UIA_LabeledByPropertyId should only return pointers to text nodes
We should only return a value for the UIA_LabeledByPropertyId when the
referenced node is a static text node. According to the spec, we should
return an empty value otherwise. This is not explicitly mentioned for
that property, but the returned value expected for that property for
each control listed in the link that can be found in the crbug is either
a static text or an empty value.

That caused an issue with alerts not being announced by Narrator on
Excel Online. From my own experimentations, Narrator appears to give
precedence to the UIA_LabeledByPropertyId value over the
UIA_NamePropertyId - i.e. when both properties have non-empty values,
we will only consider the UIA_LabeledByPropertyId.

This is what happened on Excel online: because of the structure of the
DOM tree around that alert on which there was the aria-labelledby
attribute set, the referenced node was actually the parent generic
container of the static text. Because of that, the name was computed
correctly (by AXObject::AriaTextAlternative), but the referenced node
took precedence over the name. Since the referenced node was a generic
container with an empty name, that alert went unannounced by Narrator
as it only perceived "".

Bug: 1073539
Change-Id: Ib2b5b752fd20700b80657e84d6027aed22597bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159754
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#764828}